% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanup.R
\name{cleanup}
\alias{cleanup}
\title{Census database cleanup}
\usage{
cleanup(
  df,
  year = 2017,
  tiphog = "p01",
  ocupac = "p02",
  ndorms = "p04",
  parent = "p07",
  muro = "p03a",
  techo = "p03b",
  suelo = "p03c"
)
}
\arguments{
\item{df}{data.frame object; recommendation is to use the default 'Censo 2017 Chile' database}

\item{year}{integer, self-explanatory. Supported value is 2017 (alpha version), with 2012, 2002 and 1992 coming in later revisions}

\item{tiphog}{string: tipo de hogar field. Default is 'p01'}

\item{ocupac}{string: tipo de ocupacion field. Default is 'p02'}

\item{ndorms}{string: n de dormitorios field. Default is 'p04'}

\item{parent}{string: parentesco field. Default is 'p07'}

\item{muro}{string: condicion del muro field. Default is 'p03a'}

\item{techo}{string: condicion del techo field. Default is 'p03b'}

\item{suelo}{string: condicion del suelo field. Default is 'p03c'}
}
\value{
data.frame object
}
\description{
Census database cleanup
}
\examples{
'void for now'
}
