% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/full_ismt.R
\name{full_ismt}
\alias{full_ismt}
\title{Cálculo de ISMT completo a partir del Censo 2017 -- Full ISMT calculation from 2017 Census.}
\usage{
full_ismt(
  df,
  r,
  ur,
  rfield = "region",
  urfield = "area",
  year = 2017,
  tiphog = "p01",
  ocupac = "p02",
  ndorms = "p04",
  parent = "p07",
  muro = "p03a",
  techo = "p03b",
  suelo = "p03c",
  grouping = "geocode"
)
}
\value{
data.frame with ISMT
}
\description{
'pending'
}
\examples{
ismt <- full_ismt(c17, 13, 1)
}
