% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoexport.R
\name{geoexport}
\alias{geoexport}
\title{Exportar data espacial -- Export spatial data}
\usage{
geoexport(df, r, path = "")
}
\arguments{
\item{df}{objeto \code{data.frame}. La función asume una geodata obtenida desde \code{geomerge()}. \cr \cr \code{data.frame} object. Assumes a geodata obtained from \code{geomerge()}.}

\item{r}{integer. Número de la región de trabajo. Usado para nombrar el archivo. \cr \cr integer. Number of the working region. Used to name the file.}

\item{path}{string. Ubicación de la carpeta donde se deberían guardar los archivos. Si no se provee una ubicación, se crea la carpeta \code{C:/ISMT_Files/geodata} y se utiliza como \code{path}. \cr \cr string/ Location of the directory where files should be saved. If empty, creates new directory \code{C:/ISMT_Files/geodata}, which is then used as \code{path}.}
}
\value{
Retorna el archivo en la carpeta especificada. \cr \cr Returns file in specified folder.
}
\description{
Exporta la data espacial derivada del cálculo del ISMT y posterior ejecución de \code{geomerge()} para espacializar. \cr \cr Exports spatial data derived from the calculation of ISMT and the execution of \code{geomerge()}.
}
\examples{
c17geomerge \%>\% geoexport(13, loc_dir)
}
