% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geomerge.R
\name{geomerge}
\alias{geomerge}
\title{Espacializando data del ISMT -- Spacializing ISMT data}
\usage{
geomerge(df, shp, grouping.df = "zona", grouping.shp = "GEOCODIGO")
}
\arguments{
\item{df}{objeto \code{data.frame}. Asume que los datos estarán agrupados a la unidad espacial correspondiente al shapefile especificado, tras la ejecución de \code{ismt_scores()}.}

\item{shp}{objeto \code{simple feature} (ESRI shapefile). El uso esperado incluye el shapefile censal, disponible en la página oficial del Censo o a través de \code{load_shp()} para zonas urbanas y localidades rurales.}

\item{grouping.df}{string. Nombre del campo con la variable agrupadora en la tabla de datos \cr \cr string. Name of the field containing the grouping variable in the data table.}

\item{grouping.shp}{string. Nombre del campo con la variable agrupadora en el shapefile. \cr \cr string. Name of the field containing the grouping variable in the shapefile.}
}
\value{
objeto \code{data.frame} con la información espacializada, lista para exportar. \cr \cr \code{data.frame} object containing the spacialized information, ready for export.
}
\description{
Espacializa la data obtenida del cálculo del ISMT mediante la unión con un shapefile de la unidad espacial correspondiente. \cr \cr Spacializes the data obtained from ISMT caclulation through merging with a shapefile of the corresponding spatial unit.
}
\examples{
'void for now'
}
