% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_criteria.R
\name{get_criteria}
\alias{get_criteria}
\title{Get AIM criteria for PCA}
\usage{
get_criteria(year = 2017, path = "")
}
\arguments{
\item{year}{integer}

\item{path}{string}
}
\value{
data.frame
}
\description{
Get AIM criteria for PCA
}
\examples{
'void for now'
}
