% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ismt_scores.R
\name{ismt_scores}
\alias{ismt_scores}
\title{ISMT Scores: final calculations}
\usage{
ismt_scores(
  df,
  criteria,
  r,
  ismt_score = "ismt_pn",
  grouping = "geocode",
  n = "n",
  get_criteria = F
)
}
\arguments{
\item{df}{data.frame object. Assumes database has been through the previous cleanup and calculation steps (cleanup, precalc, get_pca)}

\item{criteria}{data.frame object. Criteria for statistical cuts. As for alpha version, only 2017 AIM cuts are accepted.}

\item{r}{integer. Defines working region in order to filter criteria. Accepted values are between 1:16, and 99 (for country-wide values)}

\item{ismt_score}{string. Defines the ISMT scores field as obtained from get_pca(). Default is 'ismt_pn'}

\item{grouping}{string. Defines grouping variable. Default is 'geocode', as defined in the original Censo 2017 database. As for alpha version this is the only accepted value; future revisions will accept the 'manzent' field for 2012, 2002 and 1992 databases.}

\item{n}{string. Name of the field, expected to be derived from running cleanup(), that has a value = 1 for every home in the database. Used for counting total number of homes by grouping unit. Default is 'n'}

\item{get_criteria}{boolean. Default is FALSE. If TRUE, will temporarily download the file with the AIM criteria to C:/temp and will disappear after the execution of the function.}
}
\value{
data.frame object
}
\description{
ISMT Scores: final calculations
}
\examples{
'void for now'
}
