% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/literalize.R
\name{literalize}
\alias{literalize}
\title{Literalizar bases censales -- iteralize census databases}
\usage{
literalize(df, year)
}
\arguments{
\item{df}{objeto \code{data.frame} correspondiente a la base de datos original de los censos 20177, 2012, 2002, 1992 o 1982. -- \code{data.frame} objectcorresponding to the original database for the 2017, 2012, 2002, 1992 or 1982 census.}

\item{year}{}
}
\value{
\code{data.frame}
}
\description{
Literalizar bases censales -- iteralize census databases
}
\examples{
'no example for now'
}
