% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{load_data}
\alias{load_data}
\title{Downloading Census data}
\usage{
load_data(r, year = 2017, path = "")
}
\arguments{
\item{r}{integer. Defines the desired region. Accepted values range from 1 to 16, and 99. If r == 99, the program will download each region separately and bind them together, so it could take a minute.}

\item{year}{integer. Defines the year of the desired data. Default is 2017, which is the only valid one as for revision 0.x.x}

\item{path}{string. Defines the path of the directory where files will be stored. Default is '' (empty), which will create a new directory 'C:/ISMT_Files' and store data there.}
}
\value{
data.frame object
}
\description{
Downloading Census data
}
\examples{
'void for now'
}
