% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_shp.R
\name{load_shp}
\alias{load_shp}
\title{Load spatial information}
\usage{
load_shp(r, ur, path = "")
}
\arguments{
\item{r}{integer. Define the desired region. Acceptable values are between 1 : 16, and 99. If r == 99, country-wide data will be used.}

\item{ur}{integer. Define desired area (urban or rural). Values are 1 or 2, respectively.}

\item{path}{string. Directory where files should be downloaded. Default is '', which will create a directory to C:/ISMT_Files.}
}
\value{
simple feature object (ESRI shapefile)
}
\description{
Load spatial information
}
\examples{
'void for now'
}
