% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/namify.R
\name{namify}
\alias{namify}
\title{Asignar nombres de divisones geográficas}
\usage{
namify(df, common_var, level)
}
\arguments{
\item{df}{objeto \code{data.frame}.}

\item{common_var}{Nombre de la variable con el código común a utilizar.}

\item{level}{Nivel de la variable de referencia: r para región, p para provincia, c para comuna y zc para zona censal.}
}
\value{
df con nombres de divisiones geográficas
}
\description{
Asigna nombres de divisones geográficas en base a una variable de referencia.
}
\examples{
#''
}
