% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normvar.R
\name{normvar}
\alias{normvar}
\title{normvar: variable normalization}
\usage{
normvar(x)
}
\arguments{
\item{x}{numeric}
}
\value{
double
}
\description{
normvar: variable normalization
}
\examples{
'void for now'
}
