% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/precalc.R
\name{precalc}
\alias{precalc}
\title{Precalculations: normalized indicator scores}
\usage{
precalc(
  df,
  hacin = "ind_hacinam",
  alleg = "n_hog_alleg",
  esc = "a_esc_cont",
  mat = "ind_mater"
)
}
\arguments{
\item{df}{data.frame object. As for alpha version, assumes database has been through cleanup() function.}

\item{hacin}{string. Hacinamiento indicator field. Default is 'ind_hacinam'}

\item{alleg}{string. Allegamiento indicator field. Default is 'ind_alleg'}

\item{esc}{string. Escolaridad indicator field. Default is 'ind_esc'}

\item{mat}{string. Hacinamiento indicator field. Default is 'ind_mater'}
}
\value{
data.frame object
}
\description{
Precalculations: normalized indicator scores
}
\examples{
'void for now'
}
