% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spawn_AIMcuts.R
\name{spawn_AIMcuts}
\alias{spawn_AIMcuts}
\title{Criterio AIM para cortes estadísticos -- AIM criteria for statistical cuts}
\usage{
spawn_AIMcuts()
}
\value{
\code{data.frame} object
}
\description{
Función que no recibe parámetros, solo define un \code{data.frame} de referencia que contiene los cortes estadísticos AIM, actualizados al 2021. -- Function that does not receive any parameters, it only defines a reference \code{data.frame} containing the AIM statistical cuts, updated for 2021.
}
\examples{
ref <- spawn_AIMcuts()
}
