% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_export.R
\name{data_export}
\alias{data_export}
\title{Export ISMT data}
\usage{
data_export(df, r, path = "", format = "rds")
}
\arguments{
\item{df}{data.frame}

\item{r}{integer}

\item{path}{string}

\item{format}{string}
}
\value{
message
}
\description{
Export ISMT data
}
\examples{
'void for now'
}
