% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoexport.R
\name{geoexport}
\alias{geoexport}
\title{Export geodata}
\usage{
geoexport(df, r, path = "")
}
\arguments{
\item{df}{data.frame. Expects processed shapefile data frame after geomerge()}

\item{r}{integer. Working region}

\item{path}{string. Path where files should be saved. Default is '', which creates C:/ISMT_Files/geodata}
}
\value{
message
}
\description{
Export geodata
}
\examples{
'void for now'
}
