% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geomerge.R
\name{geomerge}
\alias{geomerge}
\title{Adding ISMT data to shapefile}
\usage{
geomerge(df, shp, grouping.df = "zona", grouping.shp = "GEOCODIGO")
}
\arguments{
\item{df}{data.frame object, corresponding to the data obtained from the previous steps. Assumes 'df' will be grouped by the appropriate geographic unit; in case of Censo2017, the grouping variable is 'geocode', which through the ismt_scores() function has been renamed 'zona'.}

\item{shp}{simple feature object (ESRI shapefile). Usage is intended for the shapefile downloaded through load_shp(), or, alternatively, the official shapefile obtained directly from INE. If downloaded directly from INE, make sure the geographic unit matches the grouping used in the calculation process.}

\item{grouping.df}{string. Name of the field containing the grouping variable in the data table.}

\item{grouping.shp}{string. Name of the field containing the grouping variable in the shapefile.}
}
\value{
data.frame object.
}
\description{
Adding ISMT data to shapefile
}
\examples{
'void for now'
}
