% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_criteria.R
\name{get_criteria}
\alias{get_criteria}
\title{Obtener criterio AIM para análisis de componentes principales -- Get AIM criteria for principal components analysis.}
\usage{
get_criteria(year = 2017, path = "")
}
\arguments{
\item{year}{integer. Default es \code{2017}, que es el único año válido para la versión \code{1.x.x}. \cr \cr integer. Default is \code{2017}, which is the only valid year for version \code{1.x.x}.}

\item{path}{string. Ubicación de la carpeta donde se deberían guardar los archivos. En caso de no proveer una ubicación, se creará un carpeta \code{C:/ISMT_Files/data}, que se usará como \code{path}. \cr \cr string. Location of the directory where files should be saved. If empty, will create a \code{C:/ISMT_Files/data} folder, which will be used as \code{path}.}
}
\value{
objeto \code{data.frame} con los valores del criterio AIM para APC. \cr \cr \code{data.frame} object with the values for the AIM criteria for PCA.
}
\description{
Carga el criterio de corte estadístico AIM para el análisis de componentes principales. Descarga el archivo con la tabla de referencia a \code{path} para futuro uso. \cr \cr Loads AIM criteria for statistical cuts for principal components analysis. Downloads file with reference table to \code{path} for future use.
}
\examples{
crit_AIM <- get_criteria(13, path = loc_dir)
}
