% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pca.R
\name{get_pca}
\alias{get_pca}
\title{Get Principal Compnents Analysis values}
\usage{
get_pca(
  df,
  esc = "ptje_esc",
  hacin = "ptje_hacin",
  mat = "ptje_mater",
  alleg = "ptje_alleg"
)
}
\arguments{
\item{df}{data.frame object. Assumes 'precalc()' function has already been run for the database, as it requires the scaled score values.}

\item{esc}{string. Escolaridad score field as obtained from precalc(). Default is 'ptje_esc'}

\item{hacin}{string. Hacinamiento score field as obtained from precalc(). Default is 'ptje_hacin'}

\item{mat}{string. Materialidad score field as obtained from precalc(). Default is 'ptje_mater'}

\item{alleg}{string. Allegamiento score field as obtained from precalc(). Default is 'ptje_alleg'}
}
\value{
data.frame object
}
\description{
Get Principal Compnents Analysis values
}
\examples{
'void for now'
}
