% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_shp.R
\name{load_shp}
\alias{load_shp}
\title{Descargar información espacial -- Download spatial information}
\usage{
load_shp(r, ur, path = "")
}
\arguments{
\item{r}{integer. Número de la región de trabajo. Acepta valores entre 1 y 16; si \code{r = 99}, se utilizan valores a nivel nacional. \cr \cr integer. Number of the working region. Accepts values between 1 and 16; if \code{r = 99}, national-level values will be used.}

\item{ur}{integer. Valores aceptables son \code{1} y \code{2}. Define si se requiere zona urbana \code{ur = 1} o rural \code{ur = 2}. \cr \cr integer. Accepted values are \code{1} and \code{2}. Defines whether urban \code{ur = 1} or rural \code{ur = 2} data is requested.}

\item{path}{string. Ubicación de la carpeta donde se deberían guardar los archivos. En caso de no proveer una ubicación, se creará un carpeta \code{C:/ISMT_Files/data}, que se usará como \code{path}. \cr \cr string. Location of the directory where files should be saved. If empty, will create a \code{C:/ISMT_Files/data} folder, which will be used as \code{path}.}
}
\value{
objeto \code{single feature}. ESRI shapefile de la región especificada. \cr \cr \code{single feature} object. ESRI shapefile of the specified region.
}
\description{
Descarga ESRI shapefile de la región especificada a \code{path} y lo carga a RStudio. \cr \cr Downloads ESRI shapefile of the specified region to \code{path} and loads it to RStudio.
}
\examples{
c17geo <- load_shp(13, 1, loc_dir)
}
