% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/region_filter.R
\name{region_filter}
\alias{region_filter}
\title{'Get only working region data'}
\usage{
region_filter(df, r, ur, rfield = "region", urfield = "area")
}
\arguments{
\item{df}{'data.frame object. As for package version 0.0.1 (alpha), only the 2017 Chilean census is accepted. Pre-formatted data is available through data(c17_chile_clean). Full data is available through data(c17_chile_full).'}

\item{r}{'integer. Values between 1 and 16 are acceptable for Chile 2017. If r == 99, no region will be defined and work will continue with national level values.'}

\item{ur}{'integer. Accepted values are 1 and 2. Defines whether urban or rural data is requested.'}

\item{rfield}{string. Name of the field corresponding to the region number. Default is 'region'}

\item{urfield}{string. Name of the field corresponding to the desired area (urban or rural). Default is 'area'}
}
\value{
data.frame object
}
\description{
'Get only working region data'
}
\examples{
'void for now'
}
